Map Child Component
Created  a function that can convert the time from string format to Date and get the hours. Once we have the hours, convert from 24hr format to AM/PM format.

In addition to this, a Child Functional Component  is created using the JS Map function, for each of the 12 hour forecast.